+++ extras/src/main/scala/japgolly/scalajs/react/extras/SetInterval.scala
@@ -1,4 +1,4 @@
-package japgolly.scalajs.react.experiment
+package japgolly.scalajs.react.extras
 
 import scala.scalajs.js
 import org.scalajs.dom.window