+++ project/Build.scala
@@ -87,7 +87,7 @@ object ScalajsReact extends Build {
 
   // ==============================================================================================
   lazy val root = Project("root", file("."))
-    .aggregate(core, test, scalaz70, scalaz71, ghpages)
+    .aggregate(core, test, scalaz70, scalaz71, extras, ghpages)
     .configure(commonSettings, preventPublication, addCommandAliases(
       "t"  -> "; test:compile ; test/fastOptStage::test",
       "tt" -> ";+test:compile ;+test/fastOptStage::test",
@@ -105,7 +105,7 @@ object ScalajsReact extends Build {
 
   lazy val test = project
     .configure(commonSettings, publicationSettings, utestSettings)
-    .dependsOn(core, scalaz71)
+    .dependsOn(core, scalaz71, extras)
     .settings(
       name := "test",
       scalacOptions += "-language:reflectiveCalls")
@@ -125,11 +125,17 @@ object ScalajsReact extends Build {
   lazy val scalaz71 = scalazModule("scalaz-7.1", "7.1.0-4")
 
   // ==============================================================================================
-  lazy val ghpages = Project("gh-pages", file("gh-pages"))
+  lazy val extras = project
+    .configure(commonSettings, publicationSettings)
     .dependsOn(core, scalaz71)
+    .settings(
+      name := "extras")
+
+  // ==============================================================================================
+  lazy val ghpages = Project("gh-pages", file("gh-pages"))
+    .dependsOn(core, scalaz71, extras)
     .configure(commonSettings, useReact(), preventPublication)
     .settings(
       emitSourceMaps := false,
       artifactPath in (Compile, fullOptJS) := file("gh-pages/res/ghpages.js"))
-
 }